+++ backend/canisters/group_index/impl/src/queries/search.rs
@@ -21,7 +21,7 @@ fn search_impl(args: Args, state: &RuntimeState) -> Response {
         return TermTooLong(MAX_TERM_LENGTH);
     }
 
-    let matches = state.data.public_groups.search(args.search_term, args.max_results);
+    let matches = state.data.public_groups.search(Some(args.search_term), 0, args.max_results);
 
     Success(SuccessResult { matches })
 }
@@ -131,6 +131,7 @@ mod tests {
                 subtype: None,
                 avatar_id: None,
                 now: env.now,
+                gate: None,
             });
         }