+++ backend/canisters/group/CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Allow blocking users who are no longer in the group ([#3818](https://github.com/open-chat-labs/open-chat/pull/3818))
+- Sync access gate with group_index ([#3826](https://github.com/open-ic/open-chat/pull/3826))
 
 ### Removed